// src/core/types/__generated__/api-schema.ts
export interface components {
    schemas: {
// ...
        PublicationResource: {
            /** @description Идентификатор публикации */
            id?: number;
            /** @description Заголовок публикации */
            title?: string;
            /** @description ЧПУ публикации */
            slug?: string;
            /** @description Изображение публикации */
            image?: string;
            category?: components["schemas"]["PublicationsCategoryResource"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}